licenta_site/pyproject.toml
[tool.poetry]
name = "licenta-site"
version = "0.1.0"
description = ""
authors = ["ignatrovino12 <ignatrovino12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.3"
gpxpy = "1.6.2"
pyexiftool = "0.5.6"
celery = "5.4.0"
django-cors-headers = "4.3.1"
google-cloud-storage = "2.16.0"
redis = "^5.0.4"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
